Vagrantfile: Work around vagrant#2281.
The Chef file cache location was changed at some point in the recent
past which exposed a race condition during initial provisioning.  The
workaround is to reset `chef.file_cache_path` to its old value.  An
alternative would be to create and share a cache from host to VM.  That
is probably a better solution and will be considered in the future.